requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiomysql==0.2.0
aioredis==2.0.0
aiosignal==1.3.1
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.9.0
azure-keyvault-secrets==4.8.0
azure-storage-queue==12.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
frozenlist==1.4.1
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
isodate==0.6.1
msal==1.31.0
msal-extensions==0.3.1
msrest==0.7.1
multidict==5.2.0
oauthlib==3.2.2
packaging==24.1
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.0.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-openapi==21.12.0
sanic-routing==23.12.0
setuptools==75.1.0
signalr-client-threads==0.0.12
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sseclient==0.0.27
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.20.0
websocket-client==1.8.0
websockets==10.0
yarl==1.13.1
